PD-2091 / 25.10 / Pd 2091 veeam smb with fast clone content (#3931)
* PD-2091 Update the Veeam Solutions Integration Article

This commit modifies the content in the /content/solutions/integrations/veeam.md article.
It adds the instructions to correctly configure data storage and the SMB share to use with Veeam with Fast Clone purpose.

* PD-2091 Update SMBPurposePresets.md Snippet

This commit updates the content in the SMB share Purpose preset snippet with current and new information on the new Purpose setting for Veeam repository with Fast Clone

* PD-2091 Update the SMB _index.md file

This commit makes grammatical changes to the content, updates section with references to Veeam options and links to the Solutions/Integrations area of the docs hub, and to the multimode share tutorial.

* PD-2091 Grammar/Spelling and Minor Content Changes

This commit makes grammatical and spelling changes to the snippet and Veeam article, and minor content changes.

* PD-2091 Fix broken link

this commit fixes a broken relref link

* Update content/SCALE/SCALETutorials/Shares/SMB/_index.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* Update content/SCALE/SCALETutorials/Shares/SMB/_index.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* Update content/SCALE/SCALETutorials/Shares/SMB/_index.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* Update content/SCALE/SCALETutorials/Shares/SMB/_index.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* Update content/SCALE/SCALETutorials/Shares/SMB/_index.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* Update content/Solutions/Integrations/Veeam.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* Update content/Solutions/Integrations/Veeam.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* PD-2091 PR Feedback Content Changes

This commit changes NFSv3 to NFSv4 in line 10, and rewrites line 172 to clarify what is written.

---------

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>