Pf 2141 update nfs shares content (#3982)
* PD-2011 update SMB Screens  UI Ref content

This commit updates the SMB share screen content with new screenshots and descriptions. Images compressed.
The SMBPorposeOtherOptions.md snippet is not in use. Might use if the tabs implemented are  recommended for removal.

* PD-2011 SMB share tutorial updates

This commit updates screenshots and content in SMB share and the multi-protocol share tutorials.

* PD-2011 Update Screenshots and External Share Sections

This commit updates a few screenshots and changes the expand section on External Shares with mixed releases to a H3 section.

* PD-2141 Update NFS Screens Content

This commit updates the NFS screen content and screenshots. Screenshots are compressed.

* PD-2011 Minor Correction

Adds internal link to line 88

* PD-2141 Minor Updates to the NFS tutorial

this commit makes minor changes and corrections to content in the AddingNFSShares.md tutorial.

* review edits

---------

Co-authored-by: Dan Pizappi <133042991+DjP-iX@users.noreply.github.com>