PD-2016 / 25.10 / Pd 2016 update network and HA network content  (#3996)
* PD-2016 Network UI Ref Updtes

This commit updates the UI ref content for Network screens. All screenshots are compressed.

* PD-2016 Update Network and HA Network Tutorials

This commit updates the network tutorials and snippets, and completes the tasks for PD-2149 by updating the HA network tutorial and snippets.

* PD-2016 Update Bridge, LAGG and VLAN Tutorials

This commit updates the tutorials for each interface type and updates screenshots. It also fixes a broken link.

* review changes

* Update GlobalConfigurationScreens.md

---------

Co-authored-by: DjP-iX <133042991+DjP-iX@users.noreply.github.com>