PD-2142 / 25.10 / Pd 2142 update i scsi UI ref and tutorial conent (#3990)
* PD-2011 update SMB Screens  UI Ref content

This commit updates the SMB share screen content with new screenshots and descriptions. Images compressed.
The SMBPorposeOtherOptions.md snippet is not in use. Might use if the tabs implemented are  recommended for removal.

* PD-2011 SMB share tutorial updates

This commit updates screenshots and content in SMB share and the multi-protocol share tutorials.

* PD-2011 Update Screenshots and External Share Sections

This commit updates a few screenshots and changes the expand section on External Shares with mixed releases to a H3 section.

* PD-2011 Minor Correction

Adds internal link to line 88

* PD-2142 Interim Commit

This commit saves iSCSI changes and screenshot added to this point. Reverted from commit to the wrong branch.

* PD-2142 iSCSI Tutorial Content Updates

This commit updates the screenshots and content in the iSCSI tutorial articles.

* PD-2142 Update Extent Associate Content

This commit updates the two snippets in the Target Extent dialog related to the remove extent association, and the associate button and dialog.

* PD-2142 Update iSCSI Screen Images

This commit updates the iSCSI targets, extents, initiators, portals, and authorized access screens, and the AddInitiatorScreen.
It removes an unnecessary carriage return in iSCSISharesScreens.md

* Update MixedModeShares.md

* Update AddingISCSIShares.md

* Update iSCSISharesScreens.md

* Update iSCSITargetDelete.md

---------

Co-authored-by: tonyriv3 <75626853+tonyriv3@users.noreply.github.com>