PD-1870 / 25.10 / Pd 1870 polish vm and container articles (#4038)
* PD-1870 Update VirtualMachinesScreens.md  Content

This commit updates the screenshots and settings content in the VirtualMachinesScreens.md UI ref article. All screenshots are compressed.

* PD-1870 Update AddManageVMDevicesSCALE.md Content

This commit updates the instructions and screenshots in the virtual machine tutorial article.

* PD-1870 Update Virtual Machine main  tutorial  _index.md file

This commit updates the content in the main /virtualMachines/_index.md file with new content and screenshots, fixes the procedure steps in ordered lists, and runs a grammar/spelling check against the content.

* Add new H4 section to UI ref

This commit adds a H4 section for the Uploading Installation Media dialog and divides the settings into separate expand sections for each of the two - screen and dialog

* PD-1870 Final Polish Edits

This commit fixes formatting and alignment issues, and missing images.

* PD-1870 Update Display Device Options

This commit updates the UI ref and tutorials to include the new **Enable Device (VNC)** option, and the **Add Device** instructions for a display.

* PD-1870 - copyedit and review

* remove identical copies

---------

Co-authored-by: Tim Moore <Mrt134@users.noreply.github.com>