PD-1882 / 25.10 / Pd 1882 update certificate articles (#4032)
* PD-1882 Update CertificateScreens.md

This commit updates the screenshots and content for the changed Certificates screen UI ref article.

* PD-1882 Update Certificates  index and tutorial content

This commit updates the Certificates screenshot and content in the Certificates/_index.md and CertificatesSCALE.md articles.

* PD-1882 Add Delete Certificate Dialog Snippet

* PD-1882 Update KMIP Content

This commit updates the KMIP tutorial, removing references to CAs, and adds new screenshots for Certificates and CSRs

* PD-1882 Updates to CSR Screens Content

This commit updates the CSRScreansSCALE.md and KMIP tutorial content with new screenshots and descriptions and instructions.

* PD-1882 Update CSR Tutorial

This commit adds new procedures to the AddCSRsSCALE.md tutorial, and makes a correction in the CertificatesScreensSCALE.md article.

* PD-1882 Final Polish to CSR Tutorial

Add screenshots to the AddCSRsSCALE.md tutorial

* PD-1882 Minor formatting change

* PD-1882 Updates to the CSR Tutorial

This commit updates the instructions in the AddCSRsSCALE.md tutorial, adds DNS warnings for the CSR when adding an ACME certificate using the CSR.

* Update AddCSRsSCALE.md

---------

Co-authored-by: Dan Pizappi <133042991+DjP-iX@users.noreply.github.com>
Co-authored-by: linzibelle <linzi.moore@ixsystems.com>
Co-authored-by: Tim Moore <Mrt134@users.noreply.github.com>
(cherry picked from commit 0d9ab50f6f986909555b8d221a1d978eab8b1663)