PD-2128 / 25.10 / PD-2128 Add User (WIP) UI Ref Article (#4041)
* PD-2128 Add User (WIP) UI Ref Article

This commit adds the screenshots and setting descriptions for the new User (WIP) screen.

A future commit will update the Additional Details settings section and screenshot after UI fixes are made to show the settings at the bottom when they are in edit mode.

* PD-2128 Compress images.

* PD-2128 User WIP UI Ref Content

This PR creates a UI ref article for the User (WIP) screen. All screenshots are compressed.
This backports to 25.10

* Update UsersScreen.md

---------

Co-authored-by: tonyriv3 <75626853+tonyriv3@users.noreply.github.com>