PD-2195 Updates to Global Network Config and Static Route Content (#4093)
This PR updates the main Network. png image, changes the global network configuration widget to the renamed NetworkConfigurationWidget.png, updated the StaticRouteWidget.png, and updated the UI ref and tutorials to reflect UI name changes. Minor changes in the StaticRoutesScreens.md UI ref. Leaving the network _index.md where it is in the main navigation for visibility rather than moving it under System