PD-2198 / 26.04 / PD-2198 Update Datasets, and Encryption Widget Info (#4095)
* PD-2198 Update Datasets, and Encryption Widget Info

This PR updates the Datasets screen and the Encryption widget content with minor changes found in NAS tickets. It updates the Encryption tutorial by changing  ZFS Encryption to just Encryption.

* PD-2198 Update the Storage Dashboard Screenshot

Tacking this screenshot update on to the Datasets ticket. This updates the screenshot of the Storage Dashboard when no pool exists/before you create a poo.

* PD-2198 Update Storage Dashboard Storage Health Widget Content

This commit changes mention of ZFS Health to Storage Health, and adds the tooltip definition of scrub to the content

* Update _index.md

---------

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>