PD-2193 Commit changes to Time Machine and Rename Branch
Discovered I used the wrong ticket when I created the branch, so it is renamed to PD-2193 in this commit.
This commit also makes changes to the SMBSharesScreens.md article adding the Enable Now option when Purpose is set to Time Machine Share and the SMB service SMB2/3 protocol option is not already enabled.