PD-2203 / 26.04 / Pd 2203 data protection UI and tutorial updates (#4100)
* PD-2203 Delete SMART Test Articles

SMART tests are removed from the UI. Deleting these articles

* PD-2203 Update Data Protection _index.md  Articles

This commit updates the content in the UI ref and tutorial Data Protection _index.md files.
Adds screenshots of the Data Protection screen before adding tasks, and compresses them. And alters the duplicated text to focus on the type of article.

* PD-2203 Save UI ref change to _index.md file

This commit saves changes in the Data Protection UI ref _index.md file.

* PD-2203 Remove Links to Deleted Articles

Remove reference links to deleted SMART and/or scrub articles in other articles.

* PD-2203 Update RsyncTaskScreens.md

This commit updates the RsyncTaskScreens.md UI ref article with new compressed screenshots and a missing setting. It includes grammar/spell check pass of the content.

* PD-2203 Update Schedule Setting Add New Snippet

This commit updates the description for the Schedule setting to work with the new snippet SchedulePresetOptions.md

and adds that snippet with expand shortcodes, and adds the AdvancedScheduler.md snippet to articles where it is missing.

* PD-2203 Update Content in Replication and Periodic Snapshot Tasks UI Ref

This commit updates screenshots, compresses existing large image files, and updates the content in the ReplicationTasksScreensSCALE.md and PeriodictSnapshotTasksScreensSCALE.md articles, runs a grammar/spell check in the TrueCloudBackupScreens.md article and the replication and periodic snapshot articles as well.

* PD-2203 Update Cloud Sync Task Content

This commit updates the widget without tasks screenshot and makes minor content changes. It makes grammar/spell checker changes.

* Update content/SCALE/SCALEUIReference/DataProtection/ReplicationScreensSCALE.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* Update content/SCALE/SCALEUIReference/DataProtection/ReplicationScreensSCALE.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* Update content/SCALE/SCALEUIReference/DataProtection/ReplicationScreensSCALE.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* PD-2203 Modify Images Replace Columns with HTML

This commit replaces the side-by-side coloumns short code with HTML code to equally size images.
It modifieds the widget images to be approximately the same size when shown side-by-side.

---------

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>