PD-2202 / 26.04 / Pd 2202 system setting updates (#4137)
* PD-2202 Upcdate General Settings Screen Content and Images

This commit updates the UI ref article for the General Settings screen with new images and content updates to match changes in the UI and flagged by the NAS tickets linked to the PD ticket.

* PD-2202 Update Advanced Settings UI Ref Delete Failover

This commit updates screenshots, adds widgets moved to the Advanced Screen, Updates content, and deletes the FailoverScreen.md article as that content moves to the Advanced Settings screen.
Grammar check completed.

* PD-2202 Update System Setting Tutorials and UI Ref

This commit updates the content in the General Settings and Advanced Settings tutorial main _index.md files and other files impacted by UI changes.
It updates screenshots (compressed) where necessary.
It also runs a grammar/spellcheck on the content files changed.

* update minor typos

---------

Co-authored-by: linzi-ix <linzi.moore@truenas.com>
Co-authored-by: linzibelle <linzi.moore@ixsystems.com>