PD-2193/ 26.04 / PD-2193 Update Legacy Purpose Setting Content (#4117)
* PD-2124 Update Legacy Purpose Setting Content

This commit modifies the settings content shown when pre-25.10 SMB shares using the No Preset with various Other Options settings selected is upgraded to 25.10. It removes the large complicated Presets comparison table as it no longer applies in 25.10..

* PD-2114 Update Tutorials

This commit updates screenshots and content in the SMB and Multi-Protocol Share Tutorials to reflect changes in the 25.10 UI.

* PD-2193 Commit changes to Time Machine and Rename Branch

Discovered I used the wrong ticket when I created the branch, so it is renamed to PD-2193 in this commit.
This commit also makes changes to the SMBSharesScreens.md article adding the Enable Now option when Purpose is set to Time Machine Share and the SMB service SMB2/3 protocol option is not already enabled.

* PD-2193 Final Edits

this commit makes the final edits to the SMBSharesScreen.md article adding the Enable Now button and description of when and where it shows. Updates the ScaleTutorials/Shares/SMB/_index.md file with a note about not using the root dataset for a share

---------

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>