PD-2238 Update Tutorials
This commit updates the ManagUsersSCALE.md tutorial, changes the organization of content in snippets, creates a new snippet for creating a new user, and modifies the API key information in the ManageUsersSCALE.md to reflect the new process and screens. The content existed but was incorrect for the new Users screen. It also fixes user content in other tutorials for admin users and in the group tutorial.