PD-2218 Update Local Replication Tutorial
This commit updates the LocalReplicationSCALE.md tutorial to use the new snippets

Renames the ReplicationWizardSteps4Thru6.md to ReplicationWizardSteps5And6.md, which moves the Replicate Custom Snapshot content up to step 4 and makes it part of the base content in the RemoteReplicationSCALE.md tutorial. It does not apply to local replication.