PD-2218 / 25.10 / Pd 2218 update replication task content (by micjohnson777) (#4204)
* Empty commit to create PR on github.

You should reset it

* PD-2218 / 26.04 / Pd 2218 update replication task content (#4191)

* PD-2218 Update Replication UI Ref

Added changes based on the way the Advanced Replication Creation button operates, and fixed a few descriptions for screens and settings that needed clarifications.

* PD-2218 Update Language about switching  between screens

This commit updates the language about switching between the wizard and add task screen to remove duplication.

* PD-2218 Replication Updates

initial changes

* PD-2218 Update UI ref and main replication tutorial index

This commit updates the ReplicationScreensSCALE.md article with new or changed definitions, and updates the BasicReplicationProcess.md snippet by breaking the local and remote overviews into individual sections, and updates each step to reflect the current processes

* PD-2218 Create New Snippets Rewrite Remote Replication

This commit rewrites the RemoteReplicationSCALE.md tutorial, using new snippets to streamline the process to set up replication tasks.

* PD-2218 Grammer, Spelling, and Content check

This commit includes a grammar/spelling check and makes a few minor changes where needed.

* PD-2218 Update Local Replication Tutorial

This commit updates the LocalReplicationSCALE.md tutorial to use the new snippets

Renames the ReplicationWizardSteps4Thru6.md to ReplicationWizardSteps5And6.md, which moves the Replicate Custom Snapshot content up to step 4 and makes it part of the base content in the RemoteReplicationSCALE.md tutorial. It does not apply to local replication.

* PD-2218 Save changes not in previous commit

* PD-2218 Saves changes in Remote Replication not in the two previous commits

* PD-2218 Update Advanced Replication

Update the AdvancedReplication/_index.md file content to work with the changed BasicReplicationProcess.md snippet.

* PD-2218 Update the ReplicationWithEncryptionSCALE.md Tutorial

This commit makes changes to the tutorial for replication using encryption. It removes the old procedure and updates with a referal to the RemoteReplicationSCALE.md tutorial, and updates that tutorial with the optional encryption setting. It removes duplicated content and focuses the article on replication and encryption options.

* PD-2218 Update the AddSSHConnection.md snippet

Added information about the connection after it is added in TrueNAS.

* typo corrections

---------

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>
Co-authored-by: linzi-ix <linzi.moore@truenas.com>
(cherry picked from commit 4c4a449532ce334a725fcd5c9f04bba4e535025f)

* PD-2218 Fix broken links

---------

Co-authored-by: MicJ <92740932+micjohnson777@users.noreply.github.com>
Co-authored-by: linzibelle <linzi.moore@ixsystems.com>
Co-authored-by: linzi-ix <linzi.moore@truenas.com>
Co-authored-by: Mic Johnson <mjohnson@ixsystems.com>