PD-1871 / 25.10 / smart content removals (25.10) (#4218)
* PD-1871: remove SMART content articles

These areas don't exist in 25.10 (or newer) content.

Added aliases to point the removed URLs towards the Advanced Settings page and add a statement to that content that smart tests from pre-25.10 systems appear as cron jobs.

Remove obsolete references to SMART services or backup tasks in the getting started content (more review to do)

* Update alert settings/hardware related alerts to account for preserved SMART alerts

* PD-1871: Finish initial file cleanup around SMART

* PD-1871: Review - search for and handle remaining textual references to SMART

SMART related settings in the Alerts area are preserved as these remain.
One SMART reference in the Advanced System Settings as a pointer for users upgrading from pre-25.10 and looking for