PD-2238 / 26.04 / Pd 2238 update new user screen content (#4160)
* PD-2238 Rename and delete old User files, fix broken Links

This commit deletes the old LocalUsersScreenSCALE.md article adds alias for this to the new UsersScreen.md file, and fixes all the broken links to the deleted article.

* PD-2238 Update Tutorials

This commit updates the ManagUsersSCALE.md tutorial, changes the organization of content in snippets, creates a new snippet for creating a new user, and modifies the API key information in the ManageUsersSCALE.md to reflect the new process and screens. The content existed but was incorrect for the new Users screen. It also fixes user content in other tutorials for admin users and in the group tutorial.

* PD-2238 Final Edits to Existing User Content

This commit makes changes to the section on adding a home directory to make it match the changes to settings.
Change the SMB user content to show the new SMB Access option instead of SMB User.

* PD-2238 Final edits to Manage User content

This commit changes the content in the Adding Home Directories section of the MsanageUsersSCALE.md article to match the new settings and process.

* PD-2238 Change Aliase

* PD-2238 Final changes to the AddingAUser.md snippet

* PD-2238 Change alias URL

* PD-2238 Fix alias in the UI ref

* review edits

* PD-2238 Update AdminRoles.md

* Update SetUpBasicTimeMachineSMBShare.md

* Update UsersScreen.md

* PD-2238 Add aliase and rename ManageUsersSCALE.md

This commit fixes broken lines after renaming ManageUsersSCALE.md to just ManageUsers.md

* PD-2238 Fix broken link

* Update RsyncTasksSCALE.md

removed close expand short code in line 60

---------

Co-authored-by: DjP-iX <133042991+DjP-iX@users.noreply.github.com>
Co-authored-by: tonyriv3 <75626853+tonyriv3@users.noreply.github.com>