PD-2187 / 26.04 / PD-2187 Update VM UI Ref and Tutorial Content (#4208)
* PD-2187 Update VM UI Ref and Tutorial Content

This PR adds information on the new functions: Export to Image and the Suspended state.
It adds missing fields to the UI ref and adds a section on exporting a disk to an image. This backports to 25.10

* review edits

---------

Co-authored-by: DjP-iX <133042991+DjP-iX@users.noreply.github.com>
(cherry picked from commit 5e9248e65b20436eb33a0aa120856f18a8d71596)