PD-2187 / 25.10 / PD-2187 Update VM UI Ref and Tutorial Content (by micjohnson777) (#4224)
* Empty commit to create PR on github.

You should reset it

* PD-2187 / 26.04 / PD-2187 Update VM UI Ref and Tutorial Content (#4208)

* PD-2187 Update VM UI Ref and Tutorial Content

This PR adds information on the new functions: Export to Image and the Suspended state.
It adds missing fields to the UI ref and adds a section on exporting a disk to an image. This backports to 25.10

* review edits

---------

Co-authored-by: DjP-iX <133042991+DjP-iX@users.noreply.github.com>
(cherry picked from commit 5e9248e65b20436eb33a0aa120856f18a8d71596)

* Update words-to-ignore.txt

* Update AddManageVMDevicesSCALE.md

* Update AddManageVMDevicesSCALE.md

---------

Co-authored-by: MicJ <92740932+micjohnson777@users.noreply.github.com>
Co-authored-by: DjP-iX <133042991+DjP-iX@users.noreply.github.com>
Co-authored-by: Mic Johnson <mjohnson@ixsystems.com>
Co-authored-by: linzibelle <linzi.moore@ixsystems.com>