Revise backup instructions for system config file (#4238)
* Revise backup instructions for system config file

Correct the section on backing up the configuration file to the original intent of focusing on easy remote backup and restore with TrueCommand.

* Rework flow of "Uploading the File" section and do Vale review

The Uploading the File section was pretty choppy with multiple admonitions and clunky phrasing. Rewrite it to flow a little better with the admonitions incorporated into the text.

* Update warning about FreeBSD config file uploads

Clarified the user account behavior when uploading a FreeBSD-based configuration file.

---------

Co-authored-by: Dan Pizappi <133042991+DjP-iX@users.noreply.github.com>