PD-2172 / 25.10 / PD-2172 Add Migrating Zvol Content (#4225)
* PD-2172 Add Migrating Zvol Content

This commit adds a section to the AddManageVMDevicesSCALE.md tutorial that covers how to move a zvol associated with an Instance VM in 25.10 and use it in a new VM.
This content only applies to the 25.10 branch, and backports to the 25.04 branch.

* PD-2127 Updates to the Tutorial and Release NOtes

This commit makes a few changes to clarify command variables in step 7, and adds a short description of the zvol migration in the 25.10 release notes.

* Update AddManageVMDevicesSCALE.md

Reword line 226.

* proofreading edits

* Update content/GettingStarted/SCALEReleaseNotes.md

Co-authored-by: Dan Pizappi <133042991+DjP-iX@users.noreply.github.com>

* PD-2172 25.10 Branch Updates

This moves the Migrating Instance VM Zvols to the /virtualmachines/_index.md file, it also updates the release notes and the main snippet file used in other related articles.

* Review Updates

* Update content/SCALETutorials/VirtualMachines/_index.md

Co-authored-by: Dan Pizappi <133042991+DjP-iX@users.noreply.github.com>

* PD-2172 PR Feedback Updates

This commit pushes changes suggested in PR feedback to clarify when users can perform this procedure and why some steps are included.

* proofreading edits

* Update words-to-ignore.txt

* Update content/SCALETutorials/VirtualMachines/_index.md

Co-authored-by: Dan Pizappi <133042991+DjP-iX@users.noreply.github.com>

* Update content/SCALETutorials/VirtualMachines/_index.md

Co-authored-by: Dan Pizappi <133042991+DjP-iX@users.noreply.github.com>

* Update content/SCALETutorials/VirtualMachines/_index.md

Co-authored-by: Dan Pizappi <133042991+DjP-iX@users.noreply.github.com>

* Update content/SCALETutorials/VirtualMachines/_index.md

Co-authored-by: Dan Pizappi <133042991+DjP-iX@users.noreply.github.com>

* Update content/SCALETutorials/VirtualMachines/_index.md

Co-authored-by: Dan Pizappi <133042991+DjP-iX@users.noreply.github.com>

* PD-2172 Feedback Updates

This commit makes changes based on suggested changes in PR feeeback

* Update words-to-ignore.txt

* Apply suggestion from @DjP-iX

* Update _index.md

---------

Co-authored-by: DjP-iX <133042991+DjP-iX@users.noreply.github.com>