PD-2172 / 25.04 / Pd 2172 add migrate instance vm content (#4231)
* PD-2172 25.04 Manual Backport of Instance VM zvol Migration

This PR manually backports the VM changes made in the 25.10 branch into the 25.04 branch. It updates the Release Notes with information about migrating Instance VM zvols and adds the procedure to the /virtuatlMachines/_index.md tutorial.

* PD-2172 Fix broken links

* PD-2172 Manual Backport of Changes

This commit makes changes to the 25.04 branch based on merged changes in 25.10 but modified for the 25.04 branch.

* PD-2172 Updates to match 25.10 Branch Changes

This commit adds information and a new command to set zvol properties to match those of natively-created VM zvols.

* Update _index.md

* Update _index.md

---------

Co-authored-by: Dan Pizappi <133042991+DjP-iX@users.noreply.github.com>
Co-authored-by: tonyriv3 <75626853+tonyriv3@users.noreply.github.com>