PD-2191 / 26.04 / PD-2191 Update the tutorial for updating an HA system (#4271)
* PD-2191 Update the tutorial for updating an HA system

This PR updates the 26.04 (nightly) branch UpdateHASCALE.md with current 25.10 UI changes related to the Update screen and process, and the content in this tutorial about updating or creating a new admin user if only the root user exists.

This backports to 25.10 branch.

* PD-2191 Add Contact Support Info

This commit adds information on contacting support for assistances.