PD-2462 Modify the Private Dataset Share Article
This PR and commit updates the SMBPrivateDatasetShare.md article with current instructions and removes instructions about home directories, which are no longer needed for private shares/directories.
It reorganizes the sections, adds a short jump-procedure to help guide the reader through the process.
It adds a section on configuring the share ACL which is necessary if you nest the private share dataset under a parent dataset.
It also adds instructions to update the parent dataset ACL to include the admin user as either owner or an entry in the ACL to prevent the error that blocks creating the private dataset share.
It updates the ManageUsers.md article, where the content under Creating A user is moved to the AddingAUser.md snippet.