PD-2462 / 26.04 / Pd 2462 update private dataset tutorial (#4347)
* PD-2462 Modify the Private Dataset Share Article

This PR and commit updates the SMBPrivateDatasetShare.md article with current instructions and removes instructions about home directories, which are no longer needed for private shares/directories.
It reorganizes the sections, adds a short jump-procedure to help guide the reader through the process.
It adds a section on configuring the share ACL which is necessary if you nest the private share dataset under a parent dataset.
It also adds instructions to update the parent dataset ACL to include the admin user as either owner or an entry in the ACL to prevent the error that blocks creating the private dataset share.
It updates the ManageUsers.md article, where the content under Creating A user is moved to the AddingAUser.md snippet.

* PD-2462 Grammar and link fixes, and minor content changes

This commit fixes broken links, corrects typos, and adds more clarification about private share user directories, and removes the first "Save Access Control List" instruction from the dataset ACL permissions section.

* PD-2462 Update Screenshots

* Update ShareACLForPrivateDatasetShare.png

* Update AddingAUser.md

* Update content/SCALE/SCALETutorials/Shares/SMB/SMBPrivateDatasetShare.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* Update SMBPrivateDatasetShare.md

change UK spelling in line 20 to US spelling

* Update SMBPrivateDatasetShare.md

Change old image short code to trueimage shortcode in line 168

---------

Co-authored-by: linzi-ix <linzi.moore@truenas.com>
Co-authored-by: linzibelle <linzi.moore@ixsystems.com>