PD-2421 / 26.04 / Pd 2421 update storage permission article in master (#4354)
* PD-2421 Content rewrites and updates of Permssions.SCALE.md article

This commit includes new screenshots, two new snippets, and reorganization/rewrites of the /datasets/permissionsSCALE.md tutorial.

* PD-2421 Update Dataset PermissionsSCALE.md Article

This PR updates the content in the PermissionsSCALE.md article with clearer explanations of POSIX and NFSv4 ACLs, instructions on how and when to  configure each. new screenshots, and an improved article organization. Two new snippets created for the common steps when adding ACE items to the ACL.

* Update content/SCALE/SCALETutorials/Datasets/PermissionsSCALE.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* Update content/SCALE/SCALETutorials/Datasets/PermissionsSCALE.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* Update PermissionsSCALE.md

Make line 98 two sentences to improve readability

* Update PermissionsSCALE.md

Change markdown bold around Permissions to html tags in line 245

---------

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>